Initial Commit
These scripts will represent a set of scripts that can manage treasury budget funds.

The treasury withdrawal should target the reward account for the `treasury` validator.

The `treasury` validator should only allow those funds to be withdrawn if they are paid out
to UTxOs locked by the `treasury` validator payment credential.

The script must also enforce that these funds are not used in the following ways:
 - Must not delegate funds to a stake pool
  - i.e. must not have a staking credential attached on the outputs
  - and must not delegate itself to a stake pool
 - Must not participate in governance
  - i.e. must not delegate to a DRep
  - must not register as a DRep itself
  - must not cast votes
  - must not submit proposals
 - Must not register a stake pool

This initial commit covers the withdrawal half of this equation (the simpler half).
Next, the spend code path will need to be built.